Locked down mutations

+++ src/resolvers/Mutation.js
@@ -1,7 +1,29 @@
 import bcrypt from "bcryptjs"
 import jwt from "jsonwebtoken"
 
+import getUserId from "../utils/getUserId"
+
 const Mutation = {
+  // AUTH
+  //
+  async login (parent, args, { prisma }, info) {
+    const user = await prisma.query.user({
+      where: {
+        email: args.data.email
+      }
+    })
+
+    if (!user) throw new Error("Unable to login!")
+
+    const isMatch = bcrypt.compare(args.data.password, user.password)
+
+    if (!isMatch) throw new Error("Unable to login!")
+
+    return {
+      user,
+      token: jwt.sign({ userId: user.id }, 'somesecret')
+    }
+  },
   // USER
   //
   async createUser (parent, args, { prisma }, info) {
@@ -9,33 +31,44 @@ const Mutation = {
 
     const password = await bcrypt.hash(args.data.password, 10)
 
-    return prisma.mutation.createUser({
+    const user = prisma.mutation.createUser({
       data: {
         ...args.data,
         password
       }
-    }, info)
+    })
+
+    return {
+      user,
+      token: jwt.sign({ userId: user.id }, 'somesecret')
+    }
   },
   //
-  deleteUser (parent, args, { prisma }, info) {
+  deleteUser (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.deleteUser({
       where: {
-        id: args.id
+        id: userId
       }
     }, info)
   },
   //
-  updateUser (parent, args, { prisma }, info) {
+  updateUser (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info)
   },
   // POST
   //
-  createPost (parent, args, { prisma }, info) {
+  createPost (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.createPost({
       data: {
         title: args.data.title,
@@ -43,14 +76,24 @@ const Mutation = {
         published: args.data.published,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         }
       }
     }, info)
   },
   //
-  deletePost (parent, args, { prisma }, info) {
+  async deletePost (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) throw new Error("Unable to delete post!")
+
     return prisma.mutation.deletePost({
       where: {
         id: args.id
@@ -58,7 +101,18 @@ const Mutation = {
     }, info)
   },
   //
-  updatePost (parent, args, { prisma }, info) {
+  async updatePost (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) throw new Error('Unable to update post!')
+
     return prisma.mutation.updatePost({
       where: {
         id: args.id
@@ -68,13 +122,15 @@ const Mutation = {
   },
   // COMMENT
   //
-  createComment (parent, args, { prisma }, info) {
+  createComment (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -86,7 +142,17 @@ const Mutation = {
     }, info)
   },
   //
-  deleteComment (parent, args, { prisma }, info) {
+  async deleteComment (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) throw new Error("Unable to delete comment")
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
@@ -94,7 +160,17 @@ const Mutation = {
     }, info)
   },
   //
-  updateComment (parent, args, { prisma }, info) {
+  async updateComment (parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) throw new Error("Unable to update comment")
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
